Formalize "finalizing" dice results a bit
Adding a third type of dice result will be necessary for things like palpatine which produces a result that cannot be modified.

Also maintained the dice count when finalizing by simply converting focuses to blanks instead of canceling them. In the majority of cases, this doesn't add any state overhead as few things conditionally add or remove dice. Accuracy corrector is the exception and interacts with lightweight frame, but since it only kicks in for 2 or less dice it's not a big performance concern.

Adding size to the state structure is expensive though, so dropping the dice counts to ubyte's and optimized the finalization process (removed multiple calls to count_all) to compensate.