Improve alpha form
Added "enable" buttons to each attacker that collapses their panels to save UI space.

Added a generic rerolls control that composes with other effects as these are common and don't want to handle the combinatoral explosion.

Added some new options to the dropdown including a few variations of lone wolf and crack shot.